+++ pkg/analysis/stdlib/testdata/distinct/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = distinct(["a", "b", "a", "c", "b"])
+}