+++ pkg/analysis/stdlib/testdata/flatten/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = flatten([["a", "b"], ["c"], ["d", "e"]])
+}