+++ pkg/analysis/stdlib/testdata/length/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = length(["a", "b", "c", "d"])
+}