+++ pkg/analysis/stdlib/testdata/tobool/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = tobool("true")
+}