+++ pkg/analysis/stdlib/testdata/tomap/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = tomap({ "k1" = "v1", "k2" = "v2" })
+}