+++ pkg/analysis/stdlib/testdata/tonumber/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = tonumber("42")
+}