+++ pkg/analysis/stdlib/testdata/tostring/main.tf
@@ -1,0 +1,3 @@
+locals {
+  out = tostring(42)
+}