+++ pkg/statediff/testdata/effective_value_collapse/length_change_still_flags/main/main.tf
@@ -1,0 +1,7 @@
+locals {
+  ids = toset(["a", "b"])
+}
+
+resource "aws_instance" "web" {
+  for_each = local.ids
+}