+++ pkg/statediff/testdata/effective_value_collapse/merge_same_value_doesnt_flag/main/main.tf
@@ -1,0 +1,7 @@
+locals {
+  tags = { "env" = "prod", "team" = "platform" }
+}
+
+resource "aws_instance" "web" {
+  for_each = local.tags
+}