+++ pkg/statediff/testdata/effective_value_collapse/sort_doesnt_flag/main/main.tf
@@ -1,0 +1,7 @@
+locals {
+  regions = ["us-east-1", "us-west-2"]
+}
+
+resource "aws_instance" "web" {
+  for_each = toset(local.regions)
+}