+++ pkg/statediff/testdata/effective_value_collapse/toset_reorder_doesnt_flag/main/main.tf
@@ -1,0 +1,7 @@
+locals {
+  ids = toset(["a", "b", "c"])
+}
+
+resource "aws_instance" "web" {
+  for_each = local.ids
+}